+++ jetpack_vendor/automattic/jetpack-external-media/src/features/admin/external-media-import.php
@@ -7,13 +7,34 @@
  * @package automattic/jetpack-external-media
  */
 
-namespace Automattic\Jetpack;
+namespace Automattic\Jetpack\External_Media;
+
+use Automattic\Jetpack\Assets;
+use Automattic\Jetpack\Connection\Manager as Connection_Manager;
+
+/**
+ * Whether the current user is connected to WordPress.com.
+ */
+function is_current_user_connected() {
+	if ( defined( 'IS_WPCOM' ) && IS_WPCOM ) {
+		return true;
+	}
+
+	return ( new Connection_Manager( 'jetpack' ) )->is_user_connected();
+}
 
 /**
  * Register the Jetpack external media page to Media > Import.
  */
 function add_jetpack_external_media_import_page() {
-	if ( empty( $_GET['jetpack_external_media_import_page'] ) ) { // phpcs:disable WordPress.Security.NonceVerification.Recommended
+	if ( empty( $_GET['jetpack_external_media_import_page'] ) && empty( $_GET['untangling-media'] ) ) { // phpcs:disable WordPress.Security.NonceVerification.Recommended
+		return;
+	}
+
+	/**
+	 * The feature is enabled only when the current user is connected to WordPress.com.
+	 */
+	if ( ! is_current_user_connected() ) {
 		return;
 	}
 
@@ -77,6 +98,7 @@ function render_jetpack_external_media_import_page() {
 	?>
 	<div id="jetpack-external-media-import" class="wrap">
 		<h1><?php echo esc_html( $title ); ?></h1>
+		<div id="jetpack-external-media-import-notice"></div>
 		<p><?php echo esc_html( $description ); ?></p>
 		<table class="widefat importers striped">
 			<?php