+++ jetpack_vendor/automattic/jetpack-external-media/src/shared/sources/google-photos/google-photos-media.js
@@ -1,6 +1,7 @@
 import { Button, SelectControl } from '@wordpress/components';
 import { useRef, useState, useCallback, useEffect } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
+import clsx from 'clsx';
 import React from 'react';
 import {
 	SOURCE_GOOGLE_PHOTOS,
@@ -27,6 +28,7 @@ const isImageOnly = allowed => allowed && allowed.length === 1 && allowed[ 0 ] =
  */
 function GooglePhotosMedia( props ) {
 	const {
+		className,
 		account,
 		allowedTypes,
 		copyMedia,
@@ -35,6 +37,7 @@ function GooglePhotosMedia( props ) {
 		isLoading,
 		media,
 		multiple,
+		selectButtonText,
 		onChangePath,
 		pageHandle,
 		path,
@@ -72,7 +75,7 @@ function GooglePhotosMedia( props ) {
 	const listUrl = getExternalMediaApiUrl( 'list', SOURCE_GOOGLE_PHOTOS, params );
 
 	const getNextPage = useCallback(
-		( event, reset = false ) => {
+		( query, reset = false ) => {
 			getMedia( listUrl, reset );
 		},
 		[ getMedia, listUrl ]
@@ -112,12 +115,12 @@ function GooglePhotosMedia( props ) {
 	useEffect( () => {
 		if ( lastQuery !== listUrl ) {
 			lastQuery.current = listUrl;
-			getNextPage( {}, path !== lastPath.current );
+			getNextPage( '', path !== lastPath.current );
 		}
 	}, [ lastQuery, listUrl, getNextPage, path ] );
 
 	return (
-		<div className="jetpack-external-media-wrapper__google">
+		<div className={ clsx( className, 'jetpack-external-media-wrapper__google' ) }>
 			<div className="jetpack-external-media-header__view">
 				{ ! pickerFeatureEnabled && (
 					<>
@@ -187,6 +190,7 @@ function GooglePhotosMedia( props ) {
 				onCopy={ onCopy }
 				pageHandle={ pageHandle }
 				multiple={ multiple }
+				selectButtonText={ selectButtonText }
 				setPath={ setPath }
 				shouldProxyImg={ pickerFeatureEnabled }
 			/>